# Project/1.PondPlanning/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-based Village Pond Planning System",
    "description": "Terrain, catchment and runoff analysis for village pond siting. Analysis routes are asynchronous: they return 202 with a job id to poll.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Poll job status",
        "description": "Return progress for one job.\n\n``stage`` is returned alongside ``progress`` because a percentage alone is\nuninformative during a 25-second pipeline; \"filling sinks\" tells the user\nsomething is happening and roughly what.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel a job",
        "description": "Request cancellation. Idempotent: cancelling a finished job is not an error.",
        "operationId": "cancel_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis/catchment": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyse Catchment",
        "description": "FR4: delineate the upstream contributing area of a clicked point.\n\nResult shape: :class:`~app.schemas.analysis.CatchmentResult`.",
        "operationId": "analyse_catchment_api_v1_analysis_catchment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatchmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runoff": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyse Runoff",
        "description": "FR6: runoff volume by three methods, reported as a range.\n\nResult shape: :class:`~app.schemas.analysis.RunoffResult`.",
        "operationId": "analyse_runoff_api_v1_analysis_runoff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/pond-design": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyse Pond Design",
        "description": "FR7: the complete pond design payload \u2014 the project's headline result.\n\nResult shape: :class:`~app.schemas.analysis.PondDesignResult`, which assembles\ncatchment, rainfall, runoff, dimensions, EAV curve, reliability, bill of\nquantities and a confidence label.",
        "operationId": "analyse_pond_design_api_v1_analysis_pond_design_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PondDesignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/suitability": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyse Suitability",
        "description": "FR3: rank candidate pond sites by AHP-weighted criteria.\n\nResult shape: :class:`~app.schemas.analysis.SuitabilityResult`.",
        "operationId": "analyse_suitability_api_v1_analysis_suitability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuitabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/results/catchment/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Catchment Result",
        "description": "FR4 result payload.",
        "operationId": "catchment_result_api_v1_analysis_results_catchment__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatchmentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/results/runoff/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Runoff Result",
        "description": "FR6 result payload.",
        "operationId": "runoff_result_api_v1_analysis_results_runoff__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunoffResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/results/pond-design/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Pond Design Result",
        "description": "FR7 result payload.",
        "operationId": "pond_design_result_api_v1_analysis_results_pond_design__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PondDesignResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/results/suitability/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Suitability Result",
        "description": "FR3 result payload.",
        "operationId": "suitability_result_api_v1_analysis_results_suitability__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuitabilityResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/results/contour/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Contour Result",
        "description": "Phase 2 result payload.",
        "operationId": "contour_result_api_v1_analysis_results_contour__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContourAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyzeContour": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyse an uploaded contour map (KML/KMZ)",
        "description": "Derive a pond location and its catchment from an uploaded contour map.\n\nEverything in the result is derived from the upload: the UTM zone from the\nfile's own centroid, the grid resolution from its own mean contour spacing,\nthe pour point from its own modelled drainage. No coordinate, extent or CRS\nspecific to any one map exists in this codebase.\n\nResult shape: :class:`~app.schemas.analysis.ContourAnalysisResult`.",
        "operationId": "analyze_contour_api_v1_analyzeContour_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_contour_api_v1_analyzeContour_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CatchmentRequest": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "pour_point": {
            "$ref": "#/components/schemas/PourPoint"
          },
          "snap_to_drainage": {
            "type": "boolean",
            "title": "Snap To Drainage",
            "description": "Move the pour point onto the nearest modelled channel before tracing upstream. Without it, a click one cell off the channel returns a hillslope catchment orders of magnitude too small.",
            "default": true
          },
          "snap_radius": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snap Radius",
            "description": "Metres. Defaults to the configured radius."
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "pour_point"
        ],
        "title": "CatchmentRequest",
        "description": "FR4: delineate the upstream contributing area of a clicked point."
      },
      "CatchmentResult": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "requested_point": {
            "$ref": "#/components/schemas/PourPoint"
          },
          "snapped_point": {
            "$ref": "#/components/schemas/PourPoint"
          },
          "snap_distance": {
            "$ref": "#/components/schemas/QuantityOut",
            "description": "How far the pour point moved. Surfaced in the UI \u2014 a large snap distance is the signal that the result should be questioned."
          },
          "area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "perimeter": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "longest_flow_path": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "mean_slope": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "relief": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "outlet_elevation": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "flow_routing": {
            "type": "string",
            "title": "Flow Routing",
            "description": "e.g. 'D8 (O'Callaghan & Mark 1984)'"
          },
          "cell_size": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "geojson": {
            "$ref": "#/components/schemas/GeoJSONFeatureCollection"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "requested_point",
          "snapped_point",
          "snap_distance",
          "area",
          "perimeter",
          "longest_flow_path",
          "mean_slope",
          "relief",
          "outlet_elevation",
          "flow_routing",
          "cell_size",
          "geojson"
        ],
        "title": "CatchmentResult",
        "description": "FR4: the delineated catchment, and enough context to trust it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url",
            "description": "GET this until status is succeeded or failed"
          },
          "estimated_seconds": {
            "type": "integer",
            "title": "Estimated Seconds",
            "description": "Rough expectation, for the progress UI"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "poll_url",
          "estimated_seconds"
        ],
        "title": "JobAccepted",
        "description": "``202`` response for every long-running analysis.\n\nTerrain analysis takes tens of seconds. Rather than hold a connection open\nand hope no proxy times it out, every analysis route returns this immediately\nand the client polls ``poll_url``."
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "Current pipeline stage, for the UI"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "progress",
          "created_at"
        ],
        "title": "JobStatus",
        "description": "Progress of an asynchronous job."
      },
...
      "PondDesignRequest": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "pour_point": {
            "$ref": "#/components/schemas/PourPoint"
          },
          "target_reliability": {
            "type": "number",
            "maximum": 0.95,
            "minimum": 0.5,
            "title": "Target Reliability",
            "description": "Fraction of years the pond should fill. 0.75 is the dependability standard used in Indian minor-irrigation practice.",
            "default": 0.75
          },
          "max_depth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Depth",
            "description": "Metres. Site or safety limit."
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "pour_point"
        ],
        "title": "PondDesignRequest",
        "description": "FR7: size a pond for a location."
      },
      "PondDesignResult": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "catchment": {
            "$ref": "#/components/schemas/CatchmentResult"
          },
          "rainfall_summary": {
            "additionalProperties": {
              "$ref": "#/components/schemas/QuantityOut"
            },
            "type": "object",
            "title": "Rainfall Summary"
          },
          "runoff": {
            "$ref": "#/components/schemas/RunoffResult"
          },
          "dimensions": {
            "$ref": "#/components/schemas/PondDimensions"
          },
          "gross_storage": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "live_storage": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "dead_storage": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "eav_curve": {
            "items": {
              "$ref": "#/components/schemas/EAVPoint"
            },
            "type": "array",
            "title": "Eav Curve"
          },
          "reliability": {
            "$ref": "#/components/schemas/QuantityOut",
            "description": "Fraction of years the pond fills"
          },
          "bill_of_quantities": {
            "$ref": "#/components/schemas/BillOfQuantities"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high"
            ],
            "title": "Confidence"
          },
          "confidence_rationale": {
            "type": "string",
            "title": "Confidence Rationale"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "catchment",
          "rainfall_summary",
          "runoff",
          "dimensions",
          "gross_storage",
          "live_storage",
          "dead_storage",
          "eav_curve",
          "reliability",
          "bill_of_quantities",
          "confidence",
          "confidence_rationale"
        ],
        "title": "PondDesignResult",
        "description": "FR7 \u2014 the full payload, and the single most important contract here.\n\nDeliberately assembled from every upstream stage rather than returned as a\nbare storage figure: catchment, rainfall, runoff, geometry, reliability,\nquantities, caveats and an overall confidence label. The confidence label is\nwhat stops a 30 m-DEM result being read as a survey."
      },
...
      "RunoffRequest": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "catchment_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Catchment Job Id",
            "description": "The catchment this runoff is computed on"
          },
          "methods": {
            "items": {
              "type": "string",
              "enum": [
                "scs_cn",
                "rational",
                "empirical_strange"
              ]
            },
            "type": "array",
            "title": "Methods",
            "description": "Runs every requested method so the answer is a range, not a point."
          },
          "curve_number": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Curve Number",
            "description": "Override the derived CN"
          },
          "years": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 5.0,
            "title": "Years",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "catchment_job_id"
        ],
        "title": "RunoffRequest",
        "description": "FR6: convert rainfall over a catchment into a runoff volume."
      },
      "RunoffResult": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "catchment_area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RunoffMethodResult"
            },
            "type": "array",
            "title": "Results"
          },
          "recommended": {
            "$ref": "#/components/schemas/RunoffMethodResult"
          },
          "spread_pct": {
            "$ref": "#/components/schemas/QuantityOut",
            "description": "Disagreement between the methods"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "catchment_area",
          "results",
          "recommended",
          "spread_pct"
        ],
        "title": "RunoffResult",
        "description": "FR6: three methods, reported as a range.\n\nThree methods rather than one because they disagree, and the disagreement is\ninformation. A single number here would be false precision."
      },
...
      "SuitabilityRequest": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "top_n": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top N",
            "default": 10
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights",
            "description": "AHP criterion weights. Derived if omitted."
          }
        },
        "type": "object",
        "required": [
          "village_id"
        ],
        "title": "SuitabilityRequest",
        "description": "FR3/P4: rank candidate pond sites across a village."
      },
...
    }
  }
}